Try to unflake BatchedFeaturedAppActivityMarkerIntegrationTest
[ci]

fixes https://github.com/DACH-NY/cn-test-failures/issues/7201 (maybe)

My best guess is there are more than 1k verdicts across tests that run
before this. The logs are useless to see anything so this both
improves the failure output as well as making sure we only look at
verdicts in a much smaller range.
